+++ Frana5u's Backgrounds/scripts/Frana5usBackgrounds/global.lua
@@ -1,4 +1,4 @@
-local U = require("scripts.Frana5usBackgrounds.utils.utils")
+local U = require("scripts.Frana5usBackgrounds.utils.scriptMerger")
 
 local folderPath = "scripts/Frana5usBackgrounds/backgrounds_global_merged/"