+++ MTR Cultures/scripts/mtrCultures/utils/raceGroups.lua
@@ -11,8 +11,9 @@ raceGroups.isKhajiit = function(player)
         ["t_els_ohmes"] = true,
         ["t_els_ohmes-raht"] = true,
         ["t_els_suthay"] = true,
+        ["t_els_tojay"] = true,
     }
-    return whitelist[playerRace]
+    return whitelist[playerRace] == true
 end
 
 raceGroups.isOrc = function(player)
@@ -22,7 +23,17 @@ raceGroups.isOrc = function(player)
         ["orc"] = true,
         ["t_mw_malahk_orc"] = true,
     }
-    return whitelist[playerRace]
+    return whitelist[playerRace] == true
+end
+
+raceGroups.isArgonian = function(player)
+    ---@diagnostic disable-next-line: undefined-field
+    local playerRace = player.type.records[player.recordId].race
+    local whitelist = {
+        ["argonian"] = true,
+        ["t_bkm_naga"] = true,
+    }
+    return whitelist[playerRace] == true
 end
 
 return raceGroups