+++ Oblivion Backgrounds/scripts/OblivionBackgrounds/global.lua
@@ -1,4 +1,4 @@
-local U = require("scripts.OblivionBackgrounds.utils.utils")
+local U = require("scripts.OblivionBackgrounds.utils.scriptMerger")
 
 local folderPath = "scripts/OblivionBackgrounds/backgrounds_global_merged/"