+++ The Wretched and the Weird/scripts/WretchedAndWeird/global.lua
@@ -1,4 +1,4 @@
-local U = require("scripts.WretchedAndWeird.utils.utils")
+local U = require("scripts.WretchedAndWeird.utils.scriptMerger")
 
 local folderPath = "scripts/WretchedAndWeird/backgrounds_global_merged/"